refactoring and tests
