+++ libs/wire-subsystems/src/Wire/EmailSubsystem/Template.hs
@@ -19,19 +19,29 @@
 
 module Wire.EmailSubsystem.Template
   ( module Wire.EmailSubsystem.Template,
+    module Wire.EmailSubsystem.Templates.Team,
+    module Wire.EmailSubsystem.Templates.User,
 
     -- * Re-exports
     Template,
+    template,
   )
 where
 
-import Data.Map qualified as Map
+import Control.Exception (catchJust)
+import Data.ByteString qualified as BS
+import Data.Map.Strict qualified as Map
+import Data.Text (pack, unpack)
+import Data.Text.Encoding qualified as T
 import Data.Text.Lazy qualified as Lazy
 import Data.Text.Template
 import HTMLEntities.Text qualified as HTML
-import Imports
+import Imports hiding (readFile)
+import System.IO.Error (isDoesNotExistError)
 import Wire.API.Locale
 import Wire.API.User
+import Wire.EmailSubsystem.Templates.Team
+import Wire.EmailSubsystem.Templates.User
 
 -- | Lookup a localised item from a 'Localised' structure.
 forLocale ::
@@ -59,8 +69,13 @@ type TemplateBranding = Text -> Text
 -- | Localised templates.
 data Localised a = Localised
   { locDefault :: (Locale, a),
-    locOther :: (Map Locale a)
+    locOther :: Map Locale a
   }
+
+-- | Uses a replace and a branding function, to replaces all placeholders from the
+-- given template to produce a Text. To be used on plain text templates
+renderText :: Template -> (Text -> Text) -> Lazy.Text
+renderText tpl replace = render tpl replace
 
 -- | Uses a replace and a branding function, to replaces all placeholders from the
 -- given template to produce a Text. To be used on plain text templates
@@ -72,157 +87,144 @@ renderTextWithBranding tpl replace branding = render tpl (replace . branding)
 renderHtmlWithBranding :: Template -> (Text -> Text) -> TemplateBranding -> Lazy.Text
 renderHtmlWithBranding tpl replace branding = render tpl (HTML.text . replace . branding)
 
-data UserTemplates = UserTemplates
-  { activationSms :: ActivationSmsTemplate,
-    activationCall :: ActivationCallTemplate,
-    verificationEmail :: VerificationEmailTemplate,
-    activationEmail :: ActivationEmailTemplate,
-    activationEmailUpdate :: ActivationEmailTemplate,
-    teamActivationEmail :: TeamActivationEmailTemplate,
-    passwordResetSms :: PasswordResetSmsTemplate,
-    passwordResetEmail :: PasswordResetEmailTemplate,
-    loginSms :: LoginSmsTemplate,
-    loginCall :: LoginCallTemplate,
-    deletionSms :: DeletionSmsTemplate,
-    deletionEmail :: DeletionEmailTemplate,
-    newClientEmail :: NewClientEmailTemplate,
-    verificationLoginEmail :: SecondFactorVerificationEmailTemplate,
-    verificationScimTokenEmail :: SecondFactorVerificationEmailTemplate,
-    verificationTeamDeletionEmail :: SecondFactorVerificationEmailTemplate
-  }
+readLocalesDir ::
+  -- | Default locale.
+  Locale ->
+  -- | Base directory.
+  FilePath ->
+  -- | Template directory (user, provider, team)
+  FilePath ->
+  -- | Handler to load the templates for a locale.
+  (FilePath -> IO a) ->
+  IO (Localised a)
+readLocalesDir defLocale base typ load = do
+  def <- load (basePath defLocaleDir)
+  Localised (defLocale, def) <$> do
+    -- Ignore locales if no such directory exist for the locale
+    ls <-
+      filterM (doesDirectoryExist . basePath)
+        . filter (/= defLocaleDir)
+        =<< listDirectory base
+    Map.fromList . zip (map readLocale ls) <$> mapM (load . basePath) ls
+  where
+    basePath :: FilePath -> FilePath
+    basePath loc = base <> "/" <> loc <> "/" <> typ
+    defLocaleDir :: FilePath
+    defLocaleDir = unpack (locToText defLocale)
+    readLocale :: FilePath -> Locale
+    readLocale l =
+      fromMaybe (error ("Invalid locale: " ++ show l)) $
+        parseLocale (pack l)
 
-data ActivationSmsTemplate = ActivationSmsTemplate
-  { activationSmslUrl :: Template,
-    activationSmsText :: Template,
-    activationSmsSender :: Text
-  }
+readTemplateWithDefault ::
+  FilePath ->
+  Locale ->
+  FilePath ->
+  String ->
+  FilePath ->
+  IO Template
+readTemplateWithDefault = readWithDefault readTemplate
 
-data ActivationCallTemplate = ActivationCallTemplate
-  { activationCallText :: Template
-  }
+readTemplate :: FilePath -> IO Template
+readTemplate f = template <$> readText f
 
-data VerificationEmailTemplate = VerificationEmailTemplate
-  { verificationEmailUrl :: Template,
-    verificationEmailSubject :: Template,
-    verificationEmailBodyText :: Template,
-    verificationEmailBodyHtml :: Template,
-    verificationEmailSender :: EmailAddress,
-    verificationEmailSenderName :: Text
-  }
+readFile :: FilePath -> IO Text
+readFile f = T.decodeUtf8 <$> BS.readFile f
 
-data ActivationEmailTemplate = ActivationEmailTemplate
-  { activationEmailUrl :: Template,
-    activationEmailSubject :: Template,
-    activationEmailBodyText :: Template,
-    activationEmailBodyHtml :: Template,
-    activationEmailSender :: EmailAddress,
-    activationEmailSenderName :: Text
-  }
+readTextWithDefault ::
+  FilePath ->
+  Locale ->
+  FilePath ->
+  String ->
+  FilePath ->
+  IO Text
+readTextWithDefault = readWithDefault readText
 
-data TeamActivationEmailTemplate = TeamActivationEmailTemplate
-  { teamActivationEmailUrl :: Template,
-    teamActivationEmailSubject :: Template,
-    teamActivationEmailBodyText :: Template,
-    teamActivationEmailBodyHtml :: Template,
-    teamActivationEmailSender :: EmailAddress,
-    teamActivationEmailSenderName :: Text
-  }
+readText :: FilePath -> IO Text
+readText f =
+  catchJust
+    (\e -> if isDoesNotExistError e then Just () else Nothing)
+    (readFile f)
+    (\_ -> error $ "Missing file: '" ++ f)
 
-data DeletionEmailTemplate = DeletionEmailTemplate
-  { deletionEmailUrl :: Template,
-    deletionEmailSubject :: Template,
-    deletionEmailBodyText :: Template,
-    deletionEmailBodyHtml :: Template,
-    deletionEmailSender :: EmailAddress,
-    deletionEmailSenderName :: Text
-  }
+readWithDefault ::
+  (String -> IO a) ->
+  FilePath ->
+  Locale ->
+  FilePath ->
+  String ->
+  FilePath ->
+  IO a
+readWithDefault readFn baseDir defLoc typ prefix name = do
+  exists <- doesFileExist fileToLoad
+  if exists
+    then readFn fileToLoad
+    else readFn fallback
+  where
+    fileToLoad = prefix <> "/" <> name
+    fallback =
+      baseDir
+        <> "/"
+        <> unpack (locToText defLoc)
+        <> "/"
+        <> typ
+        <> "/"
+        <> name
 
-data PasswordResetEmailTemplate = PasswordResetEmailTemplate
-  { passwordResetEmailUrl :: Template,
-    passwordResetEmailSubject :: Template,
-    passwordResetEmailBodyText :: Template,
-    passwordResetEmailBodyHtml :: Template,
-    passwordResetEmailSender :: EmailAddress,
-    passwordResetEmailSenderName :: Text
-  }
-
-data PasswordResetSmsTemplate = PasswordResetSmsTemplate
-  { passwordResetSmsText :: Template,
-    passwordResetSmsSender :: Text
-  }
-
-data LoginSmsTemplate = LoginSmsTemplate
-  { loginSmsUrl :: Template,
-    loginSmsText :: Template,
-    loginSmsSender :: Text
-  }
-
-data LoginCallTemplate = LoginCallTemplate
-  { loginCallText :: Template
-  }
-
-data DeletionSmsTemplate = DeletionSmsTemplate
-  { deletionSmsUrl :: Template,
-    deletionSmsText :: Template,
-    deletionSmsSender :: Text
-  }
-
-data NewClientEmailTemplate = NewClientEmailTemplate
-  { newClientEmailSubject :: Template,
-    newClientEmailBodyText :: Template,
-    newClientEmailBodyHtml :: Template,
-    newClientEmailSender :: EmailAddress,
-    newClientEmailSenderName :: Text
-  }
-
-data SecondFactorVerificationEmailTemplate = SecondFactorVerificationEmailTemplate
-  { sndFactorVerificationEmailSubject :: Template,
-    sndFactorVerificationEmailBodyText :: Template,
-    sndFactorVerificationEmailBodyHtml :: Template,
-    sndFactorVerificationEmailSender :: EmailAddress,
-    sndFactorVerificationEmailSenderName :: Text
-  }
-
-data InvitationEmailTemplate = InvitationEmailTemplate
-  { invitationEmailUrl :: !Template,
-    invitationEmailSubject :: !Template,
-    invitationEmailBodyText :: !Template,
-    invitationEmailBodyHtml :: !Template,
-    invitationEmailSender :: !EmailAddress,
-    invitationEmailSenderName :: !Text
-  }
-
-data CreatorWelcomeEmailTemplate = CreatorWelcomeEmailTemplate
-  { creatorWelcomeEmailUrl :: !Text,
-    creatorWelcomeEmailSubject :: !Template,
-    creatorWelcomeEmailBodyText :: !Template,
-    creatorWelcomeEmailBodyHtml :: !Template,
-    creatorWelcomeEmailSender :: !EmailAddress,
-    creatorWelcomeEmailSenderName :: !Text
-  }
-
-data MemberWelcomeEmailTemplate = MemberWelcomeEmailTemplate
-  { memberWelcomeEmailUrl :: !Text,
-    memberWelcomeEmailSubject :: !Template,
-    memberWelcomeEmailBodyText :: !Template,
-    memberWelcomeEmailBodyHtml :: !Template,
-    memberWelcomeEmailSender :: !EmailAddress,
-    memberWelcomeEmailSenderName :: !Text
-  }
-
-data NewTeamOwnerWelcomeEmailTemplate = NewTeamOwnerWelcomeEmailTemplate
-  { newTeamOwnerWelcomeEmailUrl :: !Text,
-    newTeamOwnerWelcomeEmailSubject :: !Template,
-    newTeamOwnerWelcomeEmailBodyText :: !Template,
-    newTeamOwnerWelcomeEmailBodyHtml :: !Template,
-    newTeamOwnerWelcomeEmailSender :: !EmailAddress,
-    newTeamOwnerWelcomeEmailSenderName :: !Text
-  }
-
-data TeamTemplates = TeamTemplates
-  { invitationEmail :: !InvitationEmailTemplate,
-    existingUserInvitationEmail :: !InvitationEmailTemplate,
-    creatorWelcomeEmail :: !CreatorWelcomeEmailTemplate,
-    memberWelcomeEmail :: !MemberWelcomeEmailTemplate,
-    newTeamOwnerWelcomeEmail :: !NewTeamOwnerWelcomeEmailTemplate
-  }
+loadTeamTemplates ::
+  Locale ->
+  FilePath ->
+  EmailAddress ->
+  Text ->
+  Text ->
+  Text ->
+  Text ->
+  IO (Localised TeamTemplates)
+loadTeamTemplates
+  defLocale
+  templateDir
+  emailSender
+  invitationEmailUrl
+  invitationEmailExistingUserUrl
+  creatorWelcomeEmailUrl
+  memberWelcomeEmailUrl =
+    readLocalesDir defLocale templateDir "team" $ \fp ->
+      TeamTemplates
+        <$> ( InvitationEmailTemplate invitationEmailUrl
+                <$> readTxt fp "email/invitation-subject.txt"
+                <*> readTxt fp "email/invitation.txt"
+                <*> readTxt fp "email/invitation.html"
+                <*> pure emailSender
+                <*> readTxt fp "email/sender.txt"
+            )
+        <*> ( InvitationEmailTemplate invitationEmailExistingUserUrl
+                <$> readTxt fp "email/migration-subject.txt"
+                <*> readTxt fp "email/migration.txt"
+                <*> readTxt fp "email/migration.html"
+                <*> pure emailSender
+                <*> readTxt fp "email/sender.txt"
+            )
+        <*> ( CreatorWelcomeEmailTemplate creatorWelcomeEmailUrl
+                <$> readTpl fp "email/new-creator-welcome-subject.txt"
+                <*> readTpl fp "email/new-creator-welcome.txt"
+                <*> readTpl fp "email/new-creator-welcome.html"
+                <*> pure emailSender
+                <*> readTxt fp "email/sender.txt"
+            )
+        <*> ( MemberWelcomeEmailTemplate memberWelcomeEmailUrl
+                <$> readTpl fp "email/new-member-welcome-subject.txt"
+                <*> readTpl fp "email/new-member-welcome.txt"
+                <*> readTpl fp "email/new-member-welcome.html"
+                <*> pure emailSender
+                <*> readTxt fp "email/sender.txt"
+            )
+        <*> ( NewTeamOwnerWelcomeEmailTemplate creatorWelcomeEmailUrl
+                <$> readTpl fp "email/new-team-owner-welcome-subject.txt"
+                <*> readTpl fp "email/new-team-owner-welcome.txt"
+                <*> readTpl fp "email/new-team-owner-welcome.html"
+                <*> pure emailSender
+                <*> readTxt fp "email/sender.txt"
+            )
+    where
+      readTpl = readTemplateWithDefault templateDir defLocale "team"
+      readTxt = readTextWithDefault templateDir defLocale "team"
