+++ services/brig/src/Brig/User/Template.hs
@@ -40,7 +40,7 @@ where
 import Brig.Options qualified as Opt
 import Brig.Template
 import Imports
-import Wire.EmailSubsystem.Template
+import Wire.EmailSubsystem.Template hiding (readTemplate, readText)
 
 loadUserTemplates :: Opt.Opts -> IO (Localised UserTemplates)
 loadUserTemplates o = readLocalesDir defLocale templateDir "user" $ \fp ->
